Move step related code
