+++ src/protocol/attribution/accumulate_credit.rs
@@ -146,7 +146,7 @@ enum Step {
     ActiveBitTimesStopBit,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
