+++ src/protocol/attribution/credit_capping.rs
@@ -465,7 +465,7 @@ enum Step {
     PrefixOrCompareBits,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
