+++ src/protocol/attribution/malicious.rs
@@ -116,7 +116,7 @@ enum Step {
     PerformUserCapping,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
