+++ src/protocol/attribution/semi_honest.rs
@@ -101,7 +101,7 @@ enum Step {
     AggregateCredit,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
