+++ src/protocol/basics/check_zero.rs
@@ -16,7 +16,7 @@ enum Step {
     RevealR,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
