+++ src/protocol/basics/mul/malicious.rs
@@ -21,7 +21,7 @@ pub(crate) enum Step {
     RandomnessForValidation,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
