+++ src/protocol/basics/sum_of_product/malicious.rs
@@ -22,7 +22,7 @@ pub(crate) enum Step {
     RandomnessForValidation,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
