+++ src/protocol/boolean/solved_bits.rs
@@ -134,7 +134,7 @@ enum Step {
     RevealC,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
