+++ src/protocol/context/malicious.rs
@@ -27,7 +27,8 @@ use crate::{
         malicious::MaliciousValidatorAccumulator,
         modulus_conversion::BitConversionTriple,
         prss::Endpoint as PrssEndpoint,
-        BitOpStep, GenericStep, NoRecord, RecordBinding, RecordId, Step,
+        step::{BitOpStep, GenericStep, Step},
+        NoRecord, RecordBinding, RecordId,
     },
     repeat64str,
     secret_sharing::{
@@ -46,7 +47,7 @@ use crate::{
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 struct UpgradeStep;
 
-impl crate::protocol::Step for UpgradeStep {}
+impl crate::protocol::step::Step for UpgradeStep {}
 
 impl AsRef<str> for UpgradeStep {
     fn as_ref(&self) -> &str {
@@ -302,7 +303,7 @@ enum UpgradeTripleStep {
     V2,
 }
 
-impl crate::protocol::Step for UpgradeTripleStep {}
+impl crate::protocol::step::Step for UpgradeTripleStep {}
 
 impl AsRef<str> for UpgradeTripleStep {
     fn as_ref(&self) -> &str {
@@ -320,7 +321,7 @@ enum UpgradeModConvStep {
     V3,
 }
 
-impl crate::protocol::Step for UpgradeModConvStep {}
+impl crate::protocol::step::Step for UpgradeModConvStep {}
 
 impl AsRef<str> for UpgradeModConvStep {
     fn as_ref(&self) -> &str {
@@ -339,7 +340,7 @@ enum UpgradeMCCappedCreditsWithAggregationBit {
     V3,
 }
 
-impl crate::protocol::Step for UpgradeMCCappedCreditsWithAggregationBit {}
+impl crate::protocol::step::Step for UpgradeMCCappedCreditsWithAggregationBit {}
 
 impl AsRef<str> for UpgradeMCCappedCreditsWithAggregationBit {
     fn as_ref(&self) -> &str {
@@ -600,7 +601,7 @@ where
 enum Upgrade2DVectors {
     V(usize),
 }
-impl crate::protocol::Step for Upgrade2DVectors {}
+impl crate::protocol::step::Step for Upgrade2DVectors {}
 
 impl AsRef<str> for Upgrade2DVectors {
     fn as_ref(&self) -> &str {
