+++ src/protocol/malicious.rs
@@ -31,7 +31,7 @@ pub(crate) enum Step {
     Validate,
 }
 
-impl crate::protocol::Step for Step {}
+impl crate::protocol::step::Step for Step {}
 
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
@@ -51,7 +51,7 @@ enum ValidateStep {
     CheckZero,
 }
 
-impl crate::protocol::Step for ValidateStep {}
+impl crate::protocol::step::Step for ValidateStep {}
 
 impl AsRef<str> for ValidateStep {
     fn as_ref(&self) -> &str {
