+++ src/protocol/sort/generate_permutation_opt.rs
@@ -13,7 +13,7 @@ use crate::{
             secureapplyinv::secureapplyinv_multi,
             SortStep::{BitPermutationStep, ComposeStep, MultiApplyInv, ShuffleRevealPermutation},
         },
-        IpaProtocolStep::Sort,
+        step::IpaProtocolStep::Sort,
     },
     secret_sharing::replicated::{
         malicious::{AdditiveShare as MaliciousReplicated, ExtendableField},
